Retry IMDS before giving up
As this pod is expected to run at early boot time, we cannot assume that
we will have an IP address yet. Normally, this would result in the pod
crashing, and kubelet retrying it. However if this happens too many
times, we enter a CrashLoopBackoff and the boostrap process is delayed.
It is better to have this gracefully wait before giving up.