Refactor manifests
The pre-existing static pods have been combined into a single "node"
manifest. This makes sense as all of the bootstrap tasks tend to be one
shots (rather than long running daemons) and many rely on each other, so
the guaranteed order that using initContainers provides makes sense. In
addition, grouping these together into a single pod removes the annoying
clutter from the api-server.

As part of this refactor, we have also added a call to the modprobe
container, which is responsible for loading three modules:

- tiny-power-button - this makes sense on AWS to handle shutdown events
  gracefully (kios-core is already updated to cope with the signal this
  module sends). This can be unloaded by cluster admins if they want to
  use the ACPI button module.

- ena - at the time of writing, this module is not separately compiled
  (it was compiled into the kernel to avoid blocking testing on AWS).
  However this is now "ready" should kios-core switch this from a Y to a
  M, as would be most sensible. It would be interesting to see what, if
  any performance ramifications this has.

- binfmt-script - in true kios fashion, we have gone all out in assuming
  what we don't know we'll need, we don't need. As a result kios-core no
  longer includes compiled in support for shebangs (#!). I am willing to
  admit that almost no admin running a cluster in AWS will expect / want
  this, however, so I will automatically reload it here, to avoid
  _everything_ being broken.

As part of this change, dhcpcd now runs as a one shot (to ensure an IP
address exists for aws-bootstrap and kubelet <-> api-server
communication. If there is a desire to run this as a long running
daemon, the recommendation is to set this up as a daemon set at the
api-server level, as this gives cluster admins most control.