Setup expected node labels
This updates aws-bootstrap to create nodes based on the AWS environment:

- topology.kubernetes.io/region
- topology.kubernetes.io/zone
- node.kubernetes.io/instance-type

Traditionally, it is kubelet's job to set these, however kiOS' kubelet
is not cloud provider aware, so it is sensible for us to set this fields
here.

NB: There are other old style labels which kubelet would normally set.
As these have been deprecated since 1.17, we will not set these:

- failure-domain.beta.kubernetes.io/region
- failure-domain.beta.kubernetes.io/zone
- beta.kubernetes.io/instance-type

Fixes #5