Move away from template-based configuration
Previously aws-bootstrap had "template" versions of configuration baked
into it, which it used as defaults. Once these were loaded and extra
dynamic / metadata-sourced configuration was applied, these would be
written to disk, clobbering the existing configuration.

We have moved away from this and now instead read the existing
configuration, using that as a base. This moves closer to the ideal of
being able to sensibly merge existing on-disk settings with
aws-bootstrap created settings.